# openapi_backend.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClaimVision-API",
    "description": "API del proyecto ClaimVision, en constante actualizacion y mejoria",
    "version": "1.12.0"
  },
  "paths": {
...
    "/api/v1/cliente/vehiculos": {
      "get": {
        "tags": [
          "v1 · Cliente"
        ],
        "summary": "Listar Mis Vehiculos",
        "operationId": "listar_mis_vehiculos_api_v1_cliente_vehiculos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_VehiculoResponseDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cliente/vehiculos/from-poliza": {
      "post": {
        "tags": [
          "v1 · Cliente"
        ],
        "summary": "Crear Vehiculo Desde Poliza",
        "description": "Crea un vehiculo a partir de los datos extraidos de una poliza de seguro via OCR.",
        "operationId": "crear_vehiculo_desde_poliza_api_v1_cliente_vehiculos_from_poliza_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_crear_vehiculo_desde_poliza_api_v1_cliente_vehiculos_from_poliza_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehiculoResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/aseguradora/crud/vehiculos": {
      "post": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Crear Vehiculo",
        "operationId": "crear_vehiculo_api_v1_aseguradora_crud_vehiculos_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehiculoResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Listar Vehiculos",
        "operationId": "listar_vehiculos_api_v1_aseguradora_crud_vehiculos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "cliente_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cliente Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_VehiculoResponseDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aseguradora/crud/vehiculos/{id}": {
      "get": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Obtener Vehiculo",
        "operationId": "obtener_vehiculo_api_v1_aseguradora_crud_vehiculos__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehiculoResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Actualizar Vehiculo",
        "operationId": "actualizar_vehiculo_api_v1_aseguradora_crud_vehiculos__id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehiculoResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Eliminar Vehiculo",
        "operationId": "eliminar_vehiculo_api_v1_aseguradora_crud_vehiculos__id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aseguradora/crud/vehiculos/from-poliza": {
      "post": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Crear Vehiculo Desde Poliza",
        "description": "Crea un vehiculo a partir de los datos extraidos de una poliza de seguro via OCR.",
        "operationId": "crear_vehiculo_desde_poliza_api_v1_aseguradora_crud_vehiculos_from_poliza_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_crear_vehiculo_desde_poliza_api_v1_aseguradora_crud_vehiculos_from_poliza_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehiculoResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_crear_vehiculo_desde_poliza_api_v1_aseguradora_crud_vehiculos_from_poliza_post": {
        "properties": {
          "cliente_id": {
            "type": "string",
            "title": "Cliente Id"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "cliente_id",
          "file"
        ],
        "title": "Body_crear_vehiculo_desde_poliza_api_v1_aseguradora_crud_vehiculos_from_poliza_post"
      },
      "Body_crear_vehiculo_desde_poliza_api_v1_cliente_vehiculos_from_poliza_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_crear_vehiculo_desde_poliza_api_v1_cliente_vehiculos_from_poliza_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_VehiculoResponseDTO_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/VehiculoResponseDTO"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[VehiculoResponseDTO]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VehiculoCreateDTO": {
        "properties": {
          "cliente_id": {
            "type": "string",
            "title": "Cliente Id"
          },
          "marca": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Marca"
          },
          "modelo": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Modelo"
          },
          "anio": {
            "type": "integer",
            "maximum": 2030.0,
            "minimum": 1900.0,
            "title": "Anio"
          },
          "placas": {
            "type": "string",
            "maxLength": 10,
            "minLength": 5,
            "title": "Placas"
          },
          "vin": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 17
              },
              {
                "type": "null"
              }
            ],
            "title": "Vin"
          },
          "color": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          }
        },
        "type": "object",
        "required": [
          "cliente_id",
          "marca",
          "modelo",
          "anio",
          "placas"
        ],
        "title": "VehiculoCreateDTO"
      },
      "VehiculoResponseDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "aseguradora_id": {
            "type": "string",
            "title": "Aseguradora Id"
          },
          "cliente_id": {
            "type": "string",
            "title": "Cliente Id"
          },
          "marca": {
            "type": "string",
            "title": "Marca"
          },
          "modelo": {
            "type": "string",
            "title": "Modelo"
          },
          "anio": {
            "type": "integer",
            "title": "Anio"
          },
          "placas": {
            "type": "string",
            "title": "Placas"
          },
          "vin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vin"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "aseguradora_id",
          "cliente_id",
          "marca",
          "modelo",
          "anio",
          "placas",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "VehiculoResponseDTO"
      },
      "VehiculoUpdateDTO": {
        "properties": {
          "marca": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marca"
          },
          "modelo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modelo"
          },
          "anio": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anio"
          },
          "placas": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placas"
          },
          "vin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vin"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          }
        },
        "type": "object",
        "title": "VehiculoUpdateDTO"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}